Initial check-in of old, but modified code.
This is the initial check in of rtbm with the old intelligence code which
depends on a closed source library removed. The goal of this new project is
just to provide a really simply way to have a Linux server give a realtime,
nicely presented and continuously updated feed of bandwidth usage.